Refactor library.
* Remove the use of the `https` library.
* Force explicit use of using a custom `httpsAgent`.
