+++ lib/VeresOne.js
@@ -3,12 +3,12 @@
  */
 'use strict';
 
-const constants = require('../constants');
-const documentLoader = require('../documentLoader');
-const VeresOneDidDoc = require('../VeresOneDidDoc');
+const constants = require('./constants');
+const documentLoader = require('./documentLoader');
+const VeresOneDidDoc = require('./VeresOneDidDoc');
 const {LDKeyPair} = require('crypto-ld');
-const VeresOneClient = require('./node/VeresOneClient');
-const VeresOneClientError = require('../VeresOneClientError');
+const VeresOneClient = require('./VeresOneClient');
+const VeresOneClientError = require('./VeresOneClientError');
 
 const jsonld = require('jsonld');
 const jsigs = require('jsonld-signatures');
@@ -21,7 +21,7 @@ class VeresOne {
    *   determines hostname for ledger client.
    * @param [options.hostname] {string} Optional hostname override. If not
    *   provided, ledger hostname will be determined based on `mode`.
-   *
+   * @param [httpsAgent] {Agent} A NodeJS HTTPS Agent (`https.Agent`) instance.
    * @param [options.logger] {object} Optional logger (defaults to console)
    * @param [options.client] {WebLedgerClient}
    *
@@ -41,7 +41,10 @@ class VeresOne {
     this.hostname = hostname;
     this.client = options.client ||
       new VeresOneClient({
-        hostname, mode: this.mode, logger: this.logger
+        hostname,
+        httpsAgent: options.httpsAgent,
+        mode: this.mode,
+        logger: this.logger
       });
 
     this.keyStore = options.keyStore;
@@ -446,7 +449,7 @@ class VeresOne {
 }
 
 VeresOne.contexts = {
-  [constants.VERES_ONE_CONTEXT_URL]: require('../contexts/veres-one-context')
+  [constants.VERES_ONE_CONTEXT_URL]: require('./contexts/veres-one-context')
 };
 
 module.exports = VeresOne;
