+++ lib/VeresOneClient.js
@@ -6,10 +6,10 @@
 const {create} = require('apisauce');
 const {WebLedgerClient} = require('web-ledger-client');
 const jsonld = require('jsonld');
-jsonld.documentLoader = require('../documentLoader');
-const VeresOneClientError = require('../VeresOneClientError');
+jsonld.documentLoader = require('./documentLoader');
+const VeresOneClientError = require('./VeresOneClientError');
 
-const constants = require('../constants');
+const constants = require('./constants');
 
 class VeresOneClient {
   /**
@@ -21,18 +21,18 @@ class VeresOneClient {
    *   agent settings (does not reject unsigned certs in 'dev' mode, for
    *   example).
    * @param [logger]
+   * @param [httpsAgent] {Agent} A NodeJS HTTPS Agent (`https.Agent`) instance.
    */
-  constructor({hostname, ledger, mode, logger}) {
+  constructor({hostname, ledger, mode, logger, httpsAgent}) {
     this.hostname = hostname;
     if(!hostname) {
       throw new Error('Missing ledger hostname.');
     }
-    const strictSSL = (mode === 'dev') ? false : true;
-    this.ledger = ledger ||
-      new WebLedgerClient({hostname, logger, strictSSL});
 
+    this.ledger = ledger || new WebLedgerClient({httpsAgent, hostname, logger});
     this.mode = mode || constants.DEFAULT_MODE;
     this.logger = logger || console;
+    this.httpsAgent = httpsAgent;
   }
 
   /**
@@ -84,7 +84,7 @@ class VeresOneClient {
     const baseURL = ticketService;
     const ticketServiceApi = create({baseURL});
     const response = await ticketServiceApi.post(
-      '/', {operation});
+      '/', {operation}, {httpsAgent: this.httpsAgent});
     if(response.problem) {
       const error = new VeresOneClientError(
         'Error retrieving record.', 'NetworkError');
@@ -146,7 +146,7 @@ class VeresOneClient {
     const baseURL = acceleratorUrl;
     const acceleratorApi = create({baseURL});
     const response = acceleratorApi.post(
-      '/', operation, {headers});
+      '/', operation, {headers, httpsAgent: this.httpsAgent});
 
     if(response.problem) {
       const error = new VeresOneClientError(
@@ -186,26 +186,25 @@ class VeresOneClient {
   }
 
   static signRequestHeaders({path, headers, signHeaders, keyId, key, method}) {
-    throw new Error('Not Implemented.');
-    // // FIXME: replace http-signature with http-signature-middleware
-    // const httpSignature = require('http-signature');
+    // FIXME: replace http-signature with http-signature-middleware
+    const httpSignature = require('http-signature');
 
-    // httpSignature.signRequest({
-    //   getHeader: header => {
-    //     // case insensitive lookup
-    //     return headers[Object.keys(headers).find(
-    //       key => key.toLowerCase() === header.toLowerCase())];
-    //   },
-    //   setHeader: (header, value) => {
-    //     headers[header] = value;
-    //   },
-    //   method,
-    //   path
-    // }, {
-    //   headers: signHeaders,
-    //   keyId,
-    //   key
-    // });
+    httpSignature.signRequest({
+      getHeader: header => {
+        // case insensitive lookup
+        return headers[Object.keys(headers).find(
+          key => key.toLowerCase() === header.toLowerCase())];
+      },
+      setHeader: (header, value) => {
+        headers[header] = value;
+      },
+      method,
+      path
+    }, {
+      headers: signHeaders,
+      keyId,
+      key
+    });
   }
 }
