+++ lib/index.js
@@ -6,7 +6,7 @@
 const constants = require('./constants');
 const documentLoader = require('./documentLoader');
 
-const VeresOne = require('./node/VeresOne');
+const VeresOne = require('./VeresOne');
 
 module.exports = {
   constants,
@@ -15,6 +15,6 @@ module.exports = {
   veres: options => {
     return new VeresOne(options);
   },
-  VeresOneClient: require('./node/VeresOneClient'),
+  VeresOneClient: require('./VeresOneClient'),
   VeresOneDidDoc: require('./VeresOneDidDoc')
 };
