+++ package.json
@@ -3,7 +3,6 @@
   "version": "6.0.1-0",
   "description": "A Decentralized Identifier utility library for Veres One",
   "main": "lib/index.js",
-  "browser": "lib/browser.js",
   "files": [
     "lib/"
   ],
